Builder: split blob and bootstrap modules from builder
Move out the building code of blob and bootstrap from the main
workflow, it makes logic clearer.
