+++ src/bin/nydus-image/core/mod.rs
@@ -2,6 +2,8 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
+pub mod blob;
+pub mod bootstrap;
 pub mod context;
 pub mod node;
 pub mod prefetch;
