+++ src/bin/nydus-image/stargz.rs
@@ -16,8 +16,6 @@ use std::rc::Rc;
 
 use nydus_utils::digest::{Algorithm, RafsDigest};
 
-pub const DEFAULT_BLOCK_SIZE: u32 = 4 << 20;
-
 type RcTocEntry = Rc<RefCell<TocEntry>>;
 
 #[derive(Deserialize, Serialize, Debug, Clone, Default)]
